fix: change method for geting position by user because of front and change constructor in domain because of lapsus calami

+++ src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/UserPosition.cs
@@ -15,7 +15,7 @@ namespace Explorer.Stakeholders.Core.Domain
         
 
 
-        public UserPosition(long userId, int latitude, int longitude)
+        public UserPosition(long userId, double latitude, double longitude)
         {
             UserId = userId;
             Latitude = latitude;
